requirements.txt
astral==3.2
pytz==2022.7.1
uvicorn
fastapi
pyswisseph